requirements.txt
Django==5.2
djangorestframework==3.17
django-filter==25.2
drf-spectacular==0.29
Faker==40
python-dotenv==1.2.2
python-slugify==8.0.4
